requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
rapidfuzz==3.11.0
pymongo==4.10.1
pyyaml==6.0.2